rag_cpgql/requirements.txt
# RAG-CPGQL System Dependencies
# Python 3.11 required

# Vector store and embeddings
chromadb>=0.4.18
sentence-transformers>=5.1.1

# Joern CPG client
cpgqls-client>=0.1.0
requests>=2.32.0

# Evaluation metrics
scikit-learn>=1.3.2
scipy>=1.11.4
numpy>=1.26.4
pandas>=2.1.3

# Data processing
jsonlines>=4.0.0
pyyaml>=6.0.1
tqdm>=4.66.1

# Utilities
python-dotenv>=1.0.0

# LLM inference - installed separately in llama.cpp conda environment
# llama-cpp-python with CUDA support should be installed manually:
# pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu124/
